+++ xnum/params.py
@@ -55,6 +55,7 @@ OSMANYA_DIGITS = "𐒠𐒡𐒢𐒣𐒤𐒥𐒦𐒧𐒨𐒩"
 MEETEI_MAYEK_DIGITS = "꯰꯱꯲꯳꯴꯵꯶꯷꯸꯹"
 KAYAH_LI_DIGITS = "꤀꤁꤂꤃꤄꤅꤆꤇꤈꤉"
 GUNJALA_GONDI_DIGITS = "𑶠𑶡𑶢𑶣𑶤𑶥𑶦𑶧𑶨𑶩"
+MASARAM_GONDI_DIGITS = "𑵐𑵑𑵒𑵓𑵔𑵕𑵖𑵗𑵘𑵙"
 
 NUMERAL_MAPS = {
     "english": ENGLISH_DIGITS,
@@ -108,6 +109,8 @@ NUMERAL_MAPS = {
     "meetei_mayek": MEETEI_MAYEK_DIGITS,
     "kayah_li": KAYAH_LI_DIGITS,
     "gunjala_gondi": GUNJALA_GONDI_DIGITS,
+    "masaram_gondi": MASARAM_GONDI_DIGITS,
+
 }
 
 ALL_DIGIT_MAPS = {}
@@ -170,6 +173,7 @@ class NumeralSystem(Enum):
     MEETEI_MAYEK = "meetei_mayek"
     KAYAH_LI = "kayah_li"
     GUNJALA_GONDI = "gunjala_gondi"
+    MASARAM_GONDI = "masaram_gondi"
     AUTO = "auto"