+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Meetei Mayek numeral system
 - Kayah Li numeral system
 - Gunjala Gondi numeral system
+- Masaram Gondi numeral system
 ## [0.9] - 2025-10-10
 ### Added
 - Dives Akuru numeral system